change datatype

+++ 9-binary_tree_height.c
@@ -7,7 +7,7 @@
  *
  * Return: size of the tree
  */
-size_t binary_tree_height(const binary_tree_t *tree)
+int binary_tree_height(const binary_tree_t *tree)
 {
 	if (!tree)
 		return (0);
@@ -15,8 +15,8 @@ size_t binary_tree_height(const binary_tree_t *tree)
 	if ((!tree->left) && (!tree->right))
 		return (0);
 
-	size_t counter_left = 0;
-	size_t counter_right = 0;
+	int counter_left = 0;
+	int counter_right = 0;
 
 	counter_left = 1 + binary_tree_height(tree->left);
 	counter_right = 1 + binary_tree_height(tree->right);
